Working on prediction model!
Implemented K-NN with the euclideian distance.
Problem is im keeping it relative to the team so it basicaly means its
traning set is ten rows(ten years of data per team)!!!
Other way is to try and work out on average how much it costs to
increase 1 position per league. more data to work off but still not
enough and doesnt take into consideration the teams previous
spending/compisition as the team might not need to spend that much as it
has already invested a ton into the team(eg. arsenal)!! i honestly just
think we dont have nearly enough data to work of.....

other way is scrape the k-nn and just write a equation to get how much
on average a team spends to gain 1 position (Probably is the best way
and looking like the only way)....

on the plus side the code i wrote is completely scalable :+1: